Add "WC_GoCardless_PayTo_Gateway_Addons" class for subscriptions and pre-orders.
