+++ includes/trait-wc-gocardless-gateway-subscriptions-pre-orders.php
@@ -1,0 +1,469 @@
+<?php
+/**
+ * Subscriptions and Pre-Orders behaviour shared by GoCardless gateways.
+ *
+ * @package WooCommerce_Gateway_GoCardless
+ */
+
+if ( ! defined( 'ABSPATH' ) ) {
+	exit;
+}
+
+/**
+ * Trait WC_GoCardless_Gateway_Subscriptions_Pre_Orders_Trait
+ *
+ * Used by {@see WC_GoCardless_Gateway_Addons} and {@see WC_GoCardless_PayTo_Gateway_Addons}.
+ */
+trait WC_GoCardless_Gateway_Subscriptions_Pre_Orders_Trait {
+
+	/**
+	 * Register WooCommerce Subscriptions and Pre-Orders hooks for this gateway ID.
+	 *
+	 * @return void
+	 */
+	protected function register_hooks() {
+		if ( class_exists( 'WC_Subscriptions_Order' ) ) {
+			add_action( 'woocommerce_scheduled_subscription_payment_' . $this->id, array( $this, 'scheduled_subscription_payment' ), 10, 2 );
+			add_action( 'woocommerce_subscription_failing_payment_method_updated_' . $this->id, array( $this, 'update_failing_payment_method' ), 10, 2 );
+
+			// Allow store managers to manually set GoCardless as the payment method on a subscription.
+			add_filter( 'woocommerce_subscription_payment_meta', array( $this, 'add_subscription_payment_meta' ), 10, 2 );
+			add_action( 'woocommerce_subscription_validate_payment_meta', array( $this, 'validate_subscription_payment_meta' ), 10, 2 );
+
+			// Cancel in-progress payment on subscription cancellation.
+			add_action( 'woocommerce_subscription_pending-cancel_' . $this->id, array( $this, 'maybe_cancel_subscription_payment' ) );
+			add_action( 'woocommerce_subscription_cancelled_' . $this->id, array( $this, 'maybe_cancel_subscription_payment' ) );
+
+			// Status synchronization for parent orders.
+			add_action( 'woocommerce_subscription_status_updated', array( $this, 'sync_parent_order_status' ), 10, 3 );
+		}
+
+		if ( class_exists( 'WC_Pre_Orders_Order' ) ) {
+			add_action( 'wc_pre_orders_process_pre_order_completion_payment_' . $this->id, array( $this, 'process_payment_for_released_pre_order' ) );
+		}
+	}
+
+	/**
+	 * Synchronize parent order status when all subscriptions are cancelled.
+	 * Also intercepts pending-cancel transitions for subscriptions with unconfirmed payments,
+	 * checking the most recent order (parent or renewal) to determine if payment is confirmed.
+	 *
+	 * @since 2.9.9
+	 * @param WC_Subscription $subscription The subscription object.
+	 * @param string          $new_status   The new subscription status.
+	 * @param string          $old_status   The old subscription status.
+	 */
+	public function sync_parent_order_status( $subscription, $new_status, $old_status ) {
+		// Only process GoCardless subscriptions and subscription cancellation triggered by the customer.
+		if ( $this->id !== $subscription->get_payment_method() || ! isset( $_GET['change_subscription_to'] ) ) { // phpcs:ignore WordPress.Security.NonceVerification.Recommended
+			return;
+		}
+
+		// Only process Some status -> 'pending-cancel' transition.
+		if ( 'pending-cancel' !== $new_status || 'pending-cancel' === $old_status ) {
+			return;
+		}
+
+		/*
+		 * Handle transition to pending-cancel status for unconfirmed payments.
+		 * This checks the most recent order's payment status (parent or latest renewal) from the
+		 * GoCardless API to avoid edge cases with webhook delays.
+		 */
+
+		// Get the most recent order of the subscription.
+		$last_order = is_callable( array( $subscription, 'get_last_order' ) )
+			? $subscription->get_last_order( 'all' )
+			: $subscription->get_parent();
+
+		// If the last order is not a valid order, return.
+		if ( ! $last_order || ! is_a( $last_order, 'WC_Abstract_Order' ) ) {
+			return;
+		}
+
+		// Get payment status from GoCardless.
+		$payment_id     = $this->get_order_resource( $last_order->get_id(), 'payment', 'id' );
+		$payment_status = '';
+
+		if ( $payment_id ) {
+			$payment = WC_GoCardless_API::get_payment( $payment_id );
+
+			if ( is_wp_error( $payment ) || empty( $payment['payments'] ) ) {
+				wc_gocardless()->log(
+					sprintf(
+						'%s - Failed to retrieve payment for order #%s',
+						__METHOD__,
+						$last_order->get_id()
+					)
+				);
+			} else {
+				$payment_status = $payment['payments']['status'] ?? '';
+			}
+		}
+
+		// Payment confirmed statuses that indicate payment has gone through.
+		$confirmed_statuses = array( 'confirmed', 'paid_out' );
+
+		// If payment is not confirmed, cancel immediately.
+		if ( ! in_array( $payment_status, $confirmed_statuses, true ) ) {
+			wc_gocardless()->log(
+				sprintf(
+					'%s - Cancelling subscription #%s immediately (order #%s has unconfirmed payment status: %s)',
+					__METHOD__,
+					$subscription->get_id(),
+					$last_order->get_id(),
+					$payment_status ? $payment_status : 'none'
+				)
+			);
+			$subscription->update_status(
+				'cancelled',
+				__( 'Subscription cancelled immediately as payment not confirmed for the last order.', 'woocommerce-gateway-gocardless' )
+			);
+		}
+	}
+
+	/**
+	 * Update GoCardless resource in order meta.
+	 *
+	 * Overrides here so that GoCardless related meta updated in subscriptions
+	 * and renewal orders.
+	 *
+	 * @since 2.4.0
+	 *
+	 * @param int|WC_Order $order         Order or Order ID.
+	 * @param string       $resource_type GoCardless resource type ('payment', 'refund' etc)
+	 *                                    in singular noun.
+	 * @param array        $resource_data Resource data.
+	 */
+	public function update_order_resource( $order, $resource_type, $resource_data = array() ) {
+		parent::update_order_resource( $order, $resource_type, $resource_data );
+
+		// Also store it on the subscriptions being purchased or paid for in the order.
+		$subscriptions = array();
+		if ( function_exists( 'wcs_order_contains_subscription' ) && wcs_order_contains_subscription( $order ) ) {
+			$subscriptions = wcs_get_subscriptions_for_order( $order );
+		} elseif ( function_exists( 'wcs_order_contains_renewal' ) && wcs_order_contains_renewal( $order ) ) {
+			$subscriptions = wcs_get_subscriptions_for_renewal_order( $order );
+		}
+
+		foreach ( $subscriptions as $subscription ) {
+			parent::update_order_resource( wc_gocardless_get_order_prop( $subscription, 'id' ), $resource_type, $resource_data );
+		}
+	}
+
+	/**
+	 * Process payment for subscription.
+	 *
+	 * @since 2.4.0
+	 *
+	 * @throws \Exception Exception.
+	 *
+	 * @param float    $amount_to_charge Amount to charge.
+	 * @param WC_Order $order            The most recent order that relates to
+	 *                                   current subscription, can be renewal
+	 *                                   order or initial order.
+	 */
+	protected function _process_subscription_payment( $amount_to_charge, $order ) {
+		$order_id   = wc_gocardless_get_order_prop( $order, 'id' );
+		$mandate_id = $this->get_order_resource( $order_id, 'mandate', 'id' );
+
+		wc_gocardless()->log( sprintf( '%s - Creating subscription payment for order #%s', __METHOD__, $order->get_order_number() ) );
+
+		add_filter( 'woocommerce_gocardless_payment_description', array( $this, 'payment_description_for_subscription' ) );
+
+		$this->_maybe_create_payment( $order_id, $mandate_id, $amount_to_charge );
+
+		remove_filter( 'woocommerce_gocardless_payment_description', array( $this, 'payment_description_for_subscription' ) );
+	}
+
+	/**
+	 * Payment description for subscription payment.
+	 *
+	 * @since 2.4.0
+	 *
+	 * @param string $description Description.
+	 *
+	 * @return string Description.
+	 */
+	public function payment_description_for_subscription( $description ) {
+		/* translators: Wrap Description with "Subscription payment from:" */
+		$description = sprintf( esc_html__( 'Subscription payment from: %s', 'woocommerce-gateway-gocardless' ), $description );
+		return $description;
+	}
+
+	/**
+	 * Process payment when a subscription payment is due.
+	 *
+	 * @since 2.4.0
+	 *
+	 * @param float    $amount_to_charge Amount to charge.
+	 * @param WC_Order $order            The most recent order that relates to
+	 *                                   current subscription, can be renewal
+	 *                                   order or initial order.
+	 */
+	public function scheduled_subscription_payment( $amount_to_charge, $order ) {
+		try {
+			$this->_process_subscription_payment( $amount_to_charge, $order );
+		} catch ( Exception $e ) {
+			/* translators: Error message */
+			$order->update_status( 'failed', sprintf( esc_html__( 'Failed to create payment via GoCardless: %s', 'woocommerce-gateway-gocardless' ), $e->getMessage() ) );
+		}
+	}
+
+	/**
+	 * Update mandate for a subscription, after using GoCardless to complete a
+	 * payment, to make up for an automatic renewal payment which previously
+	 * failed.
+	 *
+	 * @since 2.4.0
+	 *
+	 * @param WC_Subscription $subscription  The subscription for which the failing
+	 *                                       payment method relates.
+	 * @param WC_Order        $renewal_order The order which recorded the successful
+	 *                                       payment (to make up for the failed
+	 *                                       automatic payment).
+	 */
+	public function update_failing_payment_method( $subscription, $renewal_order ) {
+		$renewal_order_id = wc_gocardless_get_order_prop( $renewal_order, 'id' );
+		$this->update_order_resource( $subscription, 'mandate', $this->get_order_resource( $renewal_order_id, 'mandate' ) );
+	}
+
+	/**
+	 * Checks whether given order_id is a pre-order.
+	 *
+	 * @since 2.4.0
+	 *
+	 * @param int $order_id Order ID.
+	 *
+	 * @return boolean Returns true if pre-order.
+	 */
+	protected function _is_pre_order( $order_id ) {
+		return ( class_exists( 'WC_Pre_Orders_Order' ) && WC_Pre_Orders_Order::order_contains_pre_order( $order_id ) );
+	}
+
+	/**
+	 * Checks whether pre-order charges upon release.
+	 *
+	 * @since 2.4.0
+	 *
+	 * @param int $order_id Order ID.
+	 *
+	 * @return bool Returns true if pre-order charges upon release
+	 */
+	protected function _is_pre_order_charges_upon_release( $order_id ) {
+		return (
+			$this->_is_pre_order( $order_id )
+			&&
+			WC_Pre_Orders_Order::order_requires_payment_tokenization( $order_id )
+		);
+	}
+
+	/**
+	 * Process pre-order.
+	 *
+	 * This only process pre-order without taking payment.
+	 *
+	 * @since 2.4.0
+	 *
+	 * @param WC_Order $order Order object.
+	 *
+	 * @return array Checkout response.
+	 */
+	protected function _process_pre_order( WC_Order $order ) {
+		$order_id = wc_gocardless_get_order_prop( $order, 'id' );
+
+		wc_reduce_stock_levels( $order_id );
+
+		if ( function_exists( 'wc_empty_cart' ) ) {
+			wc_empty_cart();
+		}
+
+		WC_Pre_Orders_Order::mark_order_as_pre_ordered( $order );
+
+		return array(
+			'result'   => 'success',
+			'redirect' => $this->get_return_url( $order ),
+		);
+	}
+
+	/**
+	 * For checkout with saved token, if the order is pre-order that charges
+	 * upon release and needs tokenization then save the token in order for
+	 * taking payment later.
+	 *
+	 * @since 2.4.0
+	 *
+	 * @throws \Exception Exception.
+	 *
+	 * @param WC_Order $order Order object.
+	 */
+	protected function _process_payment_with_saved_token( WC_Order $order ) {
+		$order_id = wc_gocardless_get_order_prop( $order, 'id' );
+		if ( $this->_is_pre_order_charges_upon_release( $order_id ) ) {
+			//phpcs:disable WordPress.Security.NonceVerification.Missing -- Nonce verification is already handled on the WooCommerce side.
+			$token_input = $this->get_payment_token_input_name();
+			if ( isset( $_POST[ $token_input ] ) ) {
+				$token_id = wc_clean( wp_unslash( $_POST[ $token_input ] ) );
+			} else {
+				$token_id = false;
+			}
+			//phpcs:enable WordPress.Security.NonceVerification.Missing -- Nonce verification is already handled on the WooCommerce side.
+			$token = WC_Payment_Tokens::get( $token_id );
+			if ( ! $token || $token->get_user_id() !== get_current_user_id() ) {
+				throw new Exception( esc_html__( 'Invalid payment method. Please setup a new direct debit account.', 'woocommerce-gateway-gocardless' ) );
+			}
+
+			$mandate_id = $token->get_token();
+			$mandate    = WC_GoCardless_API::get_mandate( $mandate_id );
+			if ( is_wp_error( $mandate ) ) {
+				throw new Exception( esc_html__( 'Failed to retrieve mandate.', 'woocommerce-gateway-gocardless' ) );
+			}
+			$this->update_order_resource( $order, 'mandate', $mandate['mandates'] );
+
+			return $this->_process_pre_order( $order );
+		}
+
+		return parent::_process_payment_with_saved_token( $order );
+	}
+
+	/**
+	 * For checkout with redirect flow, and right after mandate is created,
+	 * if the order being processed is pre-order that charges upon release,
+	 * don't take payment immediately.
+	 *
+	 * @since 2.4.0
+	 *
+	 * @throws \Exception Exception.
+	 *
+	 * @param int    $order_id   Order ID.
+	 * @param string $mandate_id Mandate ID.
+	 */
+	protected function _after_mandate_created( $order_id, $mandate_id ) {
+		if ( $this->_is_pre_order_charges_upon_release( $order_id ) ) {
+			return $this->_process_pre_order( wc_get_order( $order_id ) );
+		}
+		return parent::_after_mandate_created( $order_id, $mandate_id );
+	}
+
+	/**
+	 * Process payment when pre-order is released.
+	 *
+	 * @since 2.4.0
+	 *
+	 * @param WC_Order $order Order object.
+	 */
+	public function process_payment_for_released_pre_order( $order ) {
+		$order_id   = wc_gocardless_get_order_prop( $order, 'id' );
+		$mandate_id = $this->get_order_resource( $order_id, 'mandate', 'id' );
+
+		wc_gocardless()->log( sprintf( '%s - Creating payment for pre-order #%s', __METHOD__, $order->get_order_number() ) );
+
+		try {
+			$this->_maybe_create_payment( $order_id, $mandate_id );
+		} catch ( Exception $e ) {
+			$order->update_status( 'failed' );
+			/* translators: placeholder is error message from GoCardless */
+			$order->add_order_note( sprintf( esc_html__( 'Unable to create payment: %s', 'woocommerce-gateway-gocardless' ), $e->getMessage() ) );
+		}
+	}
+
+	/**
+	 * Include the payment meta data required to process automatic recurring payments so that store managers can
+	 * manually set up automatic recurring payments for a customer via the Edit Subscriptions screen in 2.0+.
+	 *
+	 * @param array           $payment_meta associative array of meta data required for automatic payments.
+	 * @param WC_Subscription $subscription An instance of a subscription object.
+	 * @return array
+	 */
+	public function add_subscription_payment_meta( $payment_meta, $subscription ) {
+		$payment_meta[ $this->id ] = array(
+			'post_meta' => array(
+				'_gocardless_mandate_id' => array(
+					'value' => $subscription->get_meta( '_gocardless_mandate_id', true ),
+					'label' => esc_html__( 'GoCardless Mandate ID', 'woocommerce-gateway-gocardless' ),
+				),
+			),
+		);
+
+		return $payment_meta;
+	}
+
+	/**
+	 * Validate the payment meta data required to process automatic recurring payments so that store managers can
+	 * manually set up automatic recurring payments for a customer via the Edit Subscriptions screen in 2.0+.
+	 *
+	 * @param string $payment_method_id The ID of the payment method to validate.
+	 * @param array  $payment_meta      Associative array of meta data required for automatic payments.
+	 *
+	 * @throws Exception Validation Error.
+	 */
+	public function validate_subscription_payment_meta( $payment_method_id, $payment_meta ) {
+		if ( $this->id === $payment_method_id ) {
+			if ( ! isset( $payment_meta['post_meta']['_gocardless_mandate_id']['value'] ) || empty( $payment_meta['post_meta']['_gocardless_mandate_id']['value'] ) ) {
+				throw new Exception( esc_html__( 'A "GoCardless Mandate ID" value is required.', 'woocommerce-gateway-gocardless' ) );
+			}
+
+			if ( 0 !== strpos( $payment_meta['post_meta']['_gocardless_mandate_id']['value'], 'MD' ) ) {
+				throw new Exception( esc_html__( 'Invalid GoCardless Mandate ID. A valid "GoCardless Mandate ID" must begin with "MD".', 'woocommerce-gateway-gocardless' ) );
+			}
+		}
+	}
+
+	/**
+	 * Cancel the order payment on subscription cancellation.
+	 *
+	 * If the payment is pending submission, we can cancel it.
+	 * Otherwise, update the retry_if_possible to false to avoid the payment being retried (except for paid_out and cancelled payment states).
+	 *
+	 * @since 2.9.7
+	 *
+	 * @param WC_Subscription $subscription Subscription object.
+	 */
+	public function maybe_cancel_subscription_payment( $subscription ) {
+		if ( ! $subscription ) {
+			return;
+		}
+		wc_gocardless()->log( sprintf( '%s - Subscription cancelled/Pending cancellation, checking if payment should be cancelled', __METHOD__ ) );
+
+		$last_order = $subscription->get_last_order( 'all' );
+		if ( ! $last_order ) {
+			return;
+		}
+
+		$payment_id = $this->get_order_resource( $last_order->get_id(), 'payment', 'id' );
+		$payment    = WC_GoCardless_API::get_payment( $payment_id );
+		if ( is_wp_error( $payment ) || empty( $payment['payments'] ) ) {
+			wc_gocardless()->log( sprintf( '%s - Failed to retrieve payment.', __METHOD__ ) );
+			return;
+		}
+
+		$gocardless_status = $payment['payments']['status'] ?? '';
+		wc_gocardless()->log( sprintf( '%s - GoCardless payment status: %s', __METHOD__, $gocardless_status ) );
+
+		// If the payment is pending submission, we can cancel it.
+		if ( 'pending_submission' === $gocardless_status ) {
+			$response = WC_GoCardless_API::cancel_payment( $payment_id );
+			if ( ! is_wp_error( $response ) ) {
+				wc_gocardless()->log( sprintf( '%s - Payment cancelled', __METHOD__ ) );
+			} else {
+				wc_gocardless()->log( sprintf( '%s - Failed to cancel payment: %s', __METHOD__, $response->get_error_message() ) );
+			}
+		} elseif (
+			! in_array(
+				$gocardless_status,
+				array(
+					'paid_out',
+					'cancelled',
+				),
+				true
+			)
+		) {
+			// We can't cancel the payment, so we update it to not retry if possible to avoid the payment being retried.
+			$response = WC_GoCardless_API::update_payment( $payment_id, array( 'retry_if_possible' => false ) );
+			if ( ! is_wp_error( $response ) ) {
+				wc_gocardless()->log( sprintf( '%s - Payment updated with retry_if_possible set to false', __METHOD__ ) );
+			} else {
+				wc_gocardless()->log( sprintf( '%s - Failed to update payment: %s', __METHOD__, $response->get_error_message() ) );
+			}
+		}
+	}
+}
